Refactor: extract track sync logic into shared utility to fix broken server-side relative fetch (#28)
* Initial plan

* refactor: extract track sync logic into shared utility, fix server-side relative fetch


* chore: ignore package-lock.json (project uses pnpm)


---------
